Add files via upload
Reads the input file and create one Bus object per record, storing the data into an element of the Bus array.  Upon the end of the file, it displays the bus ID, manufacturer, capacity, mileage, and status for each Bus object as it was in the display so that you can be sure that you’ve built the array correctly. 

Next, read a set of input transactions from the console user.  A transaction code of ‘x’ terminates the loop. 

If the transaction code is ‘m’, it means that a maintenance record for a bus is to be added.  Ask the users for a bus ID.  Also prompt the user for the maintenance date, the description, and the name of the mechanic. If the bus exists, add the data to the maintenance history for the bus.